requirements.txt
fastapi
uvicorn
python-multipart
pydantic
google-cloud-aiplatform
google-cloud-storage
pypdf
faiss-cpu
langchain
langchain-community
langchain-text-splitters
langchain-google-vertexai
python-dotenv
